requirements.txt
dj-database-url==0.4.2
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
PyJWT==1.5.3
pytz==2017.2
whitenoise==3.3.1
